Abstracted getFifthsFromRoots to util function
